fix(checker): route lib augmentation member checks through boundary
