+++ crates/tsz-checker/src/types/queries/lib_augmentations.rs
@@ -20,50 +20,6 @@ pub(crate) fn lib_def_finalize_freeze_enabled() -> bool {
     *ENABLED.get_or_init(|| !std::env::var("TSZ_DISABLE_LIB_DEF_FREEZE").is_ok_and(|v| v == "1"))
 }
 
-/// True when `candidate`'s named-property set is a STRICT SUBSET of `current`'s
-/// — i.e. `candidate` carries every member it shares with `current` but is
-/// missing at least one that `current` has, and adds none of its own. Used to
-/// reject a heritage-thinning re-publication of a lib interface body (see
-/// [`CheckerState::register_finalized_lib_body`]).
-///
-/// Both sides must resolve to a plain object shape; if either does not (type
-/// alias, intrinsic, application, union, …) the relation is not a member-set
-/// comparison and this returns `false` so normal replacement proceeds. Disjoint
-/// or overlapping-with-additions sets also return `false`: only a pure
-/// membership loss is blocked, which keeps publication monotone in members
-/// (heritage completion and augmentation, which only grow the set, still win).
-pub(crate) fn lib_body_strictly_loses_members(
-    db: &dyn tsz_solver::construction::TypeDatabase,
-    current: TypeId,
-    candidate: TypeId,
-) -> bool {
-    if current == candidate {
-        return false;
-    }
-    let Some(current_shape) = crate::query_boundaries::common::object_shape_for_type(db, current)
-    else {
-        return false;
-    };
-    let Some(candidate_shape) =
-        crate::query_boundaries::common::object_shape_for_type(db, candidate)
-    else {
-        return false;
-    };
-    if candidate_shape.properties.len() >= current_shape.properties.len() {
-        return false;
-    }
-    // Every candidate member must already be present in `current` (no
-    // additions) and `candidate` is strictly smaller (checked above), so it is
-    // missing at least one member `current` has. Lib interface bodies carry few
-    // members, so a linear scan beats allocating a name set.
-    candidate_shape.properties.iter().all(|cand| {
-        current_shape
-            .properties
-            .iter()
-            .any(|cur| cur.name == cand.name)
-    })
-}
-
 impl<'a> CheckerState<'a> {
     /// Lower augmentation declarations from a given arena and return the resulting `TypeId`.
     ///
@@ -227,7 +183,11 @@ impl<'a> CheckerState<'a> {
             return;
         };
         self.ctx.definition_store.register_type_to_def(ty, def_id);
-        if crate::query_boundaries::common::lazy_def_id(self.ctx.types, ty) == Some(def_id) {
+        if crate::query_boundaries::lib_augmentations::is_lazy_def_identity(
+            self.ctx.types,
+            ty,
+            def_id,
+        ) {
             return;
         }
         let existing_body = self.ctx.definition_store.get_body(def_id);
@@ -264,8 +224,13 @@ impl<'a> CheckerState<'a> {
         // guards are complementary — the solver's deferred flag drops thin
         // NON-finalize re-publications, this drops thin finalize re-publications.
         let type_params = self.ctx.get_def_type_params(def_id).unwrap_or_default();
-        let keep_existing = existing_body
-            .is_some_and(|prev| lib_body_strictly_loses_members(self.ctx.types, prev, ty));
+        let keep_existing = existing_body.is_some_and(|prev| {
+            crate::query_boundaries::lib_augmentations::lib_body_strictly_loses_members(
+                self.ctx.types,
+                prev,
+                ty,
+            )
+        });
         let published = if keep_existing {
             existing_body.unwrap_or(ty)
         } else {
@@ -328,7 +293,10 @@ impl<'a> CheckerState<'a> {
         let Some(type_id) = cached else {
             return true;
         };
-        if !crate::query_boundaries::common::type_id_is_known_to_db(self.ctx.types, type_id) {
+        if !crate::query_boundaries::lib_augmentations::type_id_is_known_to_db(
+            self.ctx.types,
+            type_id,
+        ) {
             return false;
         }
         for &def_id in self.ctx.collect_lazy_def_ids_cached(type_id).iter() {
@@ -350,82 +318,9 @@ impl<'a> CheckerState<'a> {
             return true;
         }
 
-        crate::query_boundaries::common::has_construct_signatures(self.ctx.types, type_id)
-    }
-}
-
-#[cfg(test)]
-mod monotone_publication_tests {
-    //! Unit guard for the membership-monotone lib-interface body publication
-    //! that fixes the immer `SetIterator`/`MapIterator` `next`-missing false
-    //! positives (#13942): a heritage-thin re-derivation of a lib interface body
-    //! must never clobber a more-complete one in the shared store / per-file
-    //! `type_env`.
-    use super::lib_body_strictly_loses_members;
-    use crate::query_boundaries::common::{QueryDatabase, TypeInterner};
-    use tsz_solver::{PropertyInfo, TypeId};
-
-    fn obj(types: &TypeInterner, names: &[&str]) -> TypeId {
-        let props = names
-            .iter()
-            .map(|n| PropertyInfo::new(types.intern_string(n), TypeId::NUMBER))
-            .collect();
-        types.factory().object(props)
-    }
-
-    #[test]
-    fn thin_body_dropping_inherited_member_is_rejected() {
-        let types = TypeInterner::new();
-        // `SetIterator` heritage-complete: own `[Symbol.iterator]` + inherited
-        // `next` (from `IteratorObject` -> `Iterator`).
-        let complete = obj(&types, &["__@iterator", "next"]);
-        // Heritage-thin re-derivation: dropped the inherited `next`.
-        let thin = obj(&types, &["__@iterator"]);
-        assert!(
-            lib_body_strictly_loses_members(&types, complete, thin),
-            "a thin body that drops an inherited member must be rejected",
-        );
-        // Completion in the other order (thin published first, complete arriving)
-        // is a superset and must still win.
-        assert!(
-            !lib_body_strictly_loses_members(&types, thin, complete),
-            "heritage completion (growing the member set) must be allowed",
-        );
-    }
-
-    #[test]
-    fn growth_via_augmentation_is_allowed() {
-        let types = TypeInterner::new();
-        let base = obj(&types, &["a", "b"]);
-        let augmented = obj(&types, &["a", "b", "c"]);
-        assert!(!lib_body_strictly_loses_members(&types, base, augmented));
-    }
-
-    #[test]
-    fn equal_member_set_is_not_a_loss() {
-        let types = TypeInterner::new();
-        let a = obj(&types, &["a", "b"]);
-        let b = obj(&types, &["a", "b"]);
-        // Structurally identical objects intern to the same TypeId, so this also
-        // exercises the `current == candidate` short-circuit.
-        assert!(!lib_body_strictly_loses_members(&types, a, b));
-    }
-
-    #[test]
-    fn added_and_dropped_member_same_size_is_not_a_loss() {
-        let types = TypeInterner::new();
-        let current = obj(&types, &["a", "b"]);
-        // Same size, but adds `c` and drops `b`: not a strict subset, so it is
-        // not a pure membership loss and replacement proceeds.
-        let candidate = obj(&types, &["a", "c"]);
-        assert!(!lib_body_strictly_loses_members(&types, current, candidate));
-    }
-
-    #[test]
-    fn non_object_bodies_allow_replacement() {
-        let types = TypeInterner::new();
-        let o = obj(&types, &["a"]);
-        assert!(!lib_body_strictly_loses_members(&types, TypeId::NUMBER, o));
-        assert!(!lib_body_strictly_loses_members(&types, o, TypeId::STRING));
+        crate::query_boundaries::lib_augmentations::has_construct_signatures(
+            self.ctx.types,
+            type_id,
+        )
     }
 }
